mcu_code_analyzer/requirements.txt
# MCU代码分析器依赖包
# 基础依赖
pathlib2>=2.3.7
pyyaml>=6.0
requests>=2.28.0

# GUI依赖
tkinter-tooltip>=2.0.0
customtkinter>=5.2.0

# 代码分析依赖
libclang>=16.0.0
tree-sitter>=0.20.0
tree-sitter-c>=0.20.0

# 图形生成依赖
matplotlib>=3.5.0
graphviz>=0.20.0
networkx>=2.8.0

# 内嵌渲染依赖 (可选)
cefpython3>=66.0  # CEF Python for embedded browser
# pywebview>=4.0.0  # 取消注释以启用内嵌Mermaid渲染

# 本地Mermaid渲染依赖
playwright>=1.40.0  # 本地浏览器渲染引擎

# LLM集成依赖
openai>=1.0.0
anthropic>=0.7.0
ollama>=0.1.0

# 数据处理依赖
pandas>=1.5.0
numpy>=1.21.0

# 日志和配置
colorlog>=6.7.0
python-dotenv>=1.0.0

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
